pyproject.toml
[project]
name = "checkmate"
version = "0.1.0"
description = "AI-powered QA testing agent with natural language test execution"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai>=0.3.24",
    "langgraph>=0.4.8",
    "langgraph-cli[inmem]>=0.3.3",
    "langchain-mcp-adapters>=0.1.0",
    "sqlmodel>=0.0.22",
    "aiosqlite>=0.20.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "cryptography>=43.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "db", "api"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
